Split bip_coin_conf into multiple files
