+++ bip_utils/conf/common/bip_coin_conf.py
@@ -147,158 +147,3 @@ class BipCoinConf:
             AddrTypes: Address type
         """
         return self.m_addr_type
-
-
-class BipBitcoinCashConf(BipCoinConf):
-    """ Bitcoin Cash configuration class.
-    It allows to return different addresses depending on the configuration.
-    """
-
-    def __init__(self,
-                 coin_name: CoinNames,
-                 coin_idx: int,
-                 is_testnet: bool,
-                 def_path: str,
-                 key_net_ver: KeyNetVersions,
-                 wif_net_ver: bytes,
-                 bip32_type: Bip32Types,
-                 addr_conf: Dict[str, Union[bytes, str, int]],
-                 addr_type: AddrTypes,
-                 addr_type_legacy: AddrTypes) -> None:
-        """ Construct class.
-
-        Args:
-            coin_name (CoinNames object)       : Coin names
-            coin_idx (int)                     : Coin index
-            is_testnet (bool)                  : Test net flag
-            def_path (str)                     : Default path
-            key_net_ver (KeyNetVersions object): Key net versions
-            wif_net_ver (bytes)                : WIF net version
-            bip32_type (Bip32Types)            : Bip32 type
-            addr_conf (dict)                   : Address configuration
-            addr_type (AddrTypes)              : Address type
-            addr_type_legacy (AddrTypes)       : Legacy ddress type
-        """
-        super().__init__(coin_name,
-                         coin_idx,
-                         is_testnet,
-                         def_path,
-                         key_net_ver,
-                         wif_net_ver,
-                         bip32_type,
-                         addr_conf,
-                         addr_type)
-
-        self.m_addr_type_legacy = addr_type_legacy
-        self.m_use_legacy_addr = False
-
-    def UseLegacyAddress(self,
-                         value: bool) -> None:
-        """ Select if use the legacy address.
-
-        Args:
-            value (bool): True for using legacy address, false for using the standard one
-        """
-        self.m_use_legacy_addr = value
-
-    def AddrType(self) -> AddrTypes:
-        """ Get the address type. It overrides the method in BipCoinConf.
-
-        Returns:
-            AddrTypes: Address type
-        """
-        return self.m_addr_type_legacy if self.m_use_legacy_addr else self.m_addr_type
-
-    def AddrConf(self) -> Dict[str, Union[bytes, str, int]]:
-        """ Get the address configuration. It overrides the method in BipCoinConf.
-
-        Returns:
-            dict: Address configuration
-        """
-        return ({"net_ver": self.m_addr_conf["legacy_net_ver"]}
-                if self.m_use_legacy_addr
-                else {"hrp": self.m_addr_conf["std_hrp"], "net_ver": self.m_addr_conf["std_net_ver"]})
-
-
-class BipLitecoinConf(BipCoinConf):
-    """ Litecoin configuration class.
-    It allows to return different addresses and key net versions depending on the configuration.
-    """
-
-    def __init__(self,
-                 coin_name: CoinNames,
-                 coin_idx: int,
-                 is_testnet: bool,
-                 def_path: str,
-                 key_net_ver: KeyNetVersions,
-                 alt_key_net_ver: KeyNetVersions,
-                 wif_net_ver: bytes,
-                 bip32_type: Bip32Types,
-                 addr_conf: Dict[str, Union[bytes, str, int]],
-                 addr_type: AddrTypes) -> None:
-        """ Construct class.
-
-        Args:
-            coin_name (CoinNames object)           : Coin names
-            coin_idx (int)                         : Coin index
-            is_testnet (bool)                      : Test net flag
-            def_path (str)                         : Default path
-            key_net_ver (KeyNetVersions object)    : Key net versions
-            alt_key_net_ver (KeyNetVersions object): Key net versions (alternate)
-            wif_net_ver (bytes)                    : WIF net version
-            bip32_type (Bip32Types)                : Bip32 type
-            addr_conf (dict)                       : Address configuration
-            addr_type (AddrTypes)                  : Address type
-        """
-        super().__init__(coin_name,
-                         coin_idx,
-                         is_testnet,
-                         def_path,
-                         key_net_ver,
-                         wif_net_ver,
-                         bip32_type,
-                         addr_conf,
-                         addr_type)
-
-        self.m_alt_key_net_ver = alt_key_net_ver
-        self.m_use_alt_key_net_ver = False
-        self.m_use_depr_addr = False
-
-    def UseAlternateKeyNetVersions(self,
-                                   value: bool) -> None:
-        """ Select if use the alternate key net version.
-
-        Args:
-            value (bool): True for using alternate key net version, false for using the standard one
-        """
-        self.m_use_alt_key_net_ver = value
-
-    def UseDeprecatedAddress(self,
-                             value: bool) -> None:
-        """ Select if use the deprecated address.
-
-        Args:
-            value (bool): True for using deprecated address, false for using the standard one
-        """
-        self.m_use_depr_addr = value
-
-    def KeyNetVersions(self) -> KeyNetVersions:
-        """ Get key net versions. It overrides the method in BipCoinConf.
-        Litecoin overrides the method because it can have 2 different key net versions.
-
-        Returns:
-            KeyNetVersions object: KeyNetVersions object
-        """
-
-        # Get standard or alternate version depending on the configuration flag
-        return self.m_alt_key_net_ver if self.m_use_alt_key_net_ver else self.m_key_net_ver
-
-    def AddrConf(self) -> Dict[str, Union[bytes, str, int]]:
-        """ Get the address configuration. It overrides the method in BipCoinConf.
-
-        Returns:
-            dict: Address configuration
-        """
-        return ({"net_ver": self.m_addr_conf["depr_net_ver"]}
-                if self.m_use_depr_addr
-                else {"net_ver": self.m_addr_conf["std_net_ver"]})
